+++ src/fkooman/Http/Request.php
@@ -165,9 +165,14 @@ class Request
     public function setHeaders(array $requestHeaders)
     {
         foreach ($requestHeaders as $key => $value) {
-            $key = self::normalizeHeaderKey($key);
-            $this->requestHeaders[$key] = $value;
+            $this->setHeader($key, $value);
         }
+    }
+
+    public function setHeader($key, $value)
+    {
+        $key = self::normalizeHeaderKey($key);
+        $this->requestHeaders[$key] = $value;
     }
 
     public function getHeader($key)